# swagger/openapi.yaml
openapi: 3.0.3
info:
  title: Product API
  description: API for managing products and orders
  version: 1.0.0
paths:
...
  /v1/product/{product_id}:
    get:
      summary: Get product by ID
      description: Retrieves a product using its ID.
      operationId: getProductById
      parameters:
        - name: product_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Product found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '400':
          description: Invalid product ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Product not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
    Product:
      type: object
      required:
        - id
        - name
        - price
        - category
      properties:
        id:
          type: string
        name:
          type: string
        price:
          type: number
          format: float
        category:
          type: string
...
    ErrorResponse:
      type: object
      required:
        - title
        - details
      properties:
        title:   
          type: string
          example: "internal server error"
        details:
          type: string
          example: "product id should not be empty"